Added `UserPanel`
